move typeclass, datatype and extension dart files into src/
