--- lib/extension.dart
@@ -1,7 +1,0 @@
-library extension;
-
-import 'package:javelin/typeclass.dart';
-
-part 'src/extension/int.dart';
-part 'src/extension/string.dart';
-part 'src/extension/exception.dart';
