+++ lib/src/datatype/id.dart
@@ -1,4 +1,4 @@
-import 'package:javelin/typeclass.dart';
+import 'package:javelin/src/typeclass.dart';
 import 'package:javelin/src/typeclass/instance/id.dart';
 
 class ForId {
