+++ lib/src/extension.dart
@@ -1,0 +1,7 @@
+library extension;
+
+import 'package:javelin/src/typeclass.dart';
+
+part 'extension/int.dart';
+part 'extension/string.dart';
+part 'extension/exception.dart';
