+++ lib/src/typeclass.dart
@@ -1,0 +1,19 @@
+library typeclass;
+
+import 'package:javelin/src/core.dart';
+import 'package:javelin/src/datatype/either.dart';
+import 'package:javelin/src/datatype/tuple.dart';
+
+part 'hks.dart';
+part 'typeclass/applicative_error.dart';
+part 'typeclass/applicative.dart';
+part 'typeclass/apply.dart';
+part 'typeclass/bifunctor.dart';
+part 'typeclass/eq.dart';
+part 'typeclass/functor.dart';
+part 'typeclass/invariant.dart';
+part 'typeclass/monad_error.dart';
+part 'typeclass/monad.dart';
+part 'typeclass/monoid.dart';
+part 'typeclass/semigroup.dart';
+part 'typeclass/show.dart';
