+++ lib/src/typeclass/instance/either.dart
@@ -1,5 +1,5 @@
-import 'package:javelin/datatype.dart';
-import 'package:javelin/typeclass.dart';
+import 'package:javelin/src/datatype.dart';
+import 'package:javelin/src/typeclass.dart';
 
 //final eitherTypeInstance = EitherType<Object>();
