+++ lib/src/typeclass/instance/id.dart
@@ -1,4 +1,4 @@
-import 'package:javelin/typeclass.dart';
+import 'package:javelin/src/typeclass.dart';
 import 'package:javelin/src/datatype/id.dart';
 
 final idTypeInstance = IdType._();
