--- lib/typeclass.dart
@@ -1,19 +1,0 @@
-library typeclass;
-
-import 'package:javelin/src/core.dart';
-import 'package:javelin/src/datatype/either.dart';
-import 'package:javelin/src/datatype/tuple.dart';
-
-part 'src/hks.dart';
-part 'src/typeclass/applicative_error.dart';
-part 'src/typeclass/applicative.dart';
-part 'src/typeclass/apply.dart';
-part 'src/typeclass/bifunctor.dart';
-part 'src/typeclass/eq.dart';
-part 'src/typeclass/functor.dart';
-part 'src/typeclass/invariant.dart';
-part 'src/typeclass/monad_error.dart';
-part 'src/typeclass/monad.dart';
-part 'src/typeclass/monoid.dart';
-part 'src/typeclass/semigroup.dart';
-part 'src/typeclass/show.dart';
