+++ test/extension/exception_test.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:io';
 
-import 'package:javelin/extension.dart';
+import 'package:javelin/src/extension.dart';
 import 'package:test/test.dart';
 
 import '../law.dart';
