+++ test/typeclass/laws/bifunctor_laws.dart
@@ -1,4 +1,4 @@
-import 'package:javelin/typeclass.dart';
+import 'package:javelin/src/typeclass.dart';
 import 'package:javelin/src/core.dart' as core;
 
 import '../../gen.dart';
